refactor : Line 인터페이스 분리 및 필요없는 기능 제거
